+++ lib/src/core/server.dart
@@ -212,7 +212,7 @@ class Angel extends AngelBase {
     handlerCache.clear();
     _injections.clear();
     encoders.clear();
-    //_serializer = god.serialize;
+    //_serializer = json.encode;
     _children.clear();
     _parent = null;
     logger = null;