+++ lib/src/core/typed_service.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 import 'dart:mirrors';
-import 'package:json_god/json_god.dart' as god;
 import '../../common.dart';
 import 'service.dart';
 
@@ -56,7 +55,7 @@ class TypedService<T> extends Service {
 
   serialize(x) {
     if (x is Model)
-      return god.serializeObject(x);
+      return json.encodeObject(x);
     else if (x is Map)
       return x;
     else if (x is Iterable)